Rename: 내일 입찰 열리는 굿즈 관련 파일들 이름 수정
