+++ src/containers/main/ComingSoon.tsx
@@ -1,7 +1,7 @@
 import Data from '@/dummydata/comingSoon.json'
-import { ComingSoonAndRankType } from '@/types/ComingSoonAndRankType'
-import ComingSoonAndRankItem from './ComingSoonAndRankItem'
+import { SoonAndHitsType } from '@/types/SoonAndHitsType'
 import ComingSoonTitle from './ComingSoonTitle'
+import SoonAndHitsItem from './SoonAndHitsItem'
 
 export default function ComingSoon() {
   const data = Data.goodsList
@@ -10,8 +10,8 @@ export default function ComingSoon() {
     <div className="pt-[20px]">
       <ComingSoonTitle />
       <div className="pt-[10px] pb-[20px] px-[20px] grid grid-cols-2 gap-3">
-        {data.map((item: ComingSoonAndRankType) => (
-          <ComingSoonAndRankItem key={item.goodsCode} item={item} />
+        {data.map((item: SoonAndHitsType) => (
+          <SoonAndHitsItem key={item.goodsCode} item={item} />
         ))}
       </div>
     </div>
